fix: fix commands

+++ src/commands/rm.js
@@ -2,7 +2,10 @@ import { unlink } from 'node:fs/promises';
 import path from 'path';
 
 export const rm = async (currentPath, pathToFile) => {
-  const filePath = path.join(currentPath, pathToFile);
+  if (!pathToFile) {
+    throw new Error('Invalid Input');
+  }
+  const filePath = path.resolve(currentPath, pathToFile);
   try {
     await unlink(filePath);
   } catch {
